refactor: cli -> testnetwork and doc comments
