+++ app/wasmext/wasm_cli_test/cli_test.go
@@ -18,9 +18,9 @@ import (
 	"github.com/NibiruChain/nibiru/app"
 	"github.com/NibiruChain/nibiru/x/common/denoms"
 	"github.com/NibiruChain/nibiru/x/common/testutil"
-	testutilcli "github.com/NibiruChain/nibiru/x/common/testutil/cli"
 	"github.com/NibiruChain/nibiru/x/common/testutil/genesis"
 	"github.com/NibiruChain/nibiru/x/common/testutil/testapp"
+	"github.com/NibiruChain/nibiru/x/common/testutil/testnetwork"
 )
 
 // commonArgs is args for CLI test commands.
@@ -36,8 +36,8 @@ var _ suite.TearDownAllSuite = (*TestSuite)(nil)
 type TestSuite struct {
 	suite.Suite
 
-	cfg     testutilcli.Config
-	network *testutilcli.Network
+	cfg     testnetwork.Config
+	network *testnetwork.Network
 }
 
 func (s *TestSuite) SetupSuite() {
@@ -46,8 +46,8 @@ func (s *TestSuite) SetupSuite() {
 
 	encodingConfig := app.MakeEncodingConfig()
 	genesisState := genesis.NewTestGenesisState(encodingConfig)
-	s.cfg = testutilcli.BuildNetworkConfig(genesisState)
-	network, err := testutilcli.New(s.T(), s.T().TempDir(), s.cfg)
+	s.cfg = testnetwork.BuildNetworkConfig(genesisState)
+	network, err := testnetwork.New(s.T(), s.T().TempDir(), s.cfg)
 	s.Require().NoError(err)
 
 	s.network = network
@@ -96,7 +96,7 @@ func (s *TestSuite) deployWasmContract(path string) (uint64, error) {
 		return 0, err
 	}
 
-	resp, err = testutilcli.QueryTx(val.ClientCtx, resp.TxHash)
+	resp, err = testnetwork.QueryTx(val.ClientCtx, resp.TxHash)
 	if err != nil {
 		return 0, err
 	}
@@ -129,7 +129,7 @@ func (s *TestSuite) deployWasmContract(path string) (uint64, error) {
 func (s *TestSuite) requiredDeployedContractsLen(total int) {
 	val := s.network.Validators[0]
 	var queryCodeResponse types.QueryCodesResponse
-	err := testutilcli.ExecQuery(
+	err := testnetwork.ExecQuery(
 		val.ClientCtx,
 		wasmcli.GetCmdListCode(),
 		[]string{},
