--- x/common/testutil/cli/logger.go
@@ -1,14 +1,0 @@
-package cli
-
-import (
-	"testing"
-)
-
-// Logger is a network logger interface that exposes testnet-level Log() methods for an in-process testing network
-// This is not to be confused with logging that may happen at an individual node or validator level
-type Logger interface {
-	Log(args ...interface{})
-	Logf(format string, args ...interface{})
-}
-
-var _ Logger = (*testing.T)(nil)
