+++ x/common/testutil/testnetwork/network_config.go
@@ -1,4 +1,4 @@
-package cli
+package testnetwork
 
 import (
 	"encoding/json"
@@ -15,7 +15,7 @@ import (
 	serverconfig "github.com/NibiruChain/nibiru/app/server/config"
 )
 
-// Config: Defines the parameters needed to start a local test network.
+// Config: Defines the parameters needed to start a local test [Network].
 type Config struct {
 	Codec             codec.Codec
 	LegacyAmino       *codec.LegacyAmino // TODO: Remove!
