+++ x/common/testutil/testnetwork/query.go
@@ -1,4 +1,4 @@
-package cli
+package testnetwork
 
 import (
 	"fmt"
