+++ x/common/testutil/testnetwork/tx.go
@@ -1,4 +1,4 @@
-package cli
+package testnetwork
 
 import (
 	"context"
