+++ x/common/testutil/testnetwork/util.go
@@ -1,4 +1,4 @@
-package cli
+package testnetwork
 
 import (
 	"encoding/json"
